separate OdbProxy into smaller files
